feat(EconomyService): extract income, upkeep, buy-unit logic
refactor(GameService): delegate economy methods to EconomyService
