+++ src/main/kotlin/at/aau/hexabrawl/websocketserver/service/EconomyService.kt
@@ -1,0 +1,97 @@
+package at.aau.hexabrawl.websocketserver.model
+
+import org.springframework.stereotype.Service
+
+/**
+ * Wirtschafts-Logik des Spiels: Income- und Upkeep-Berechnung, Buy-Unit,
+ * Per-Player-Economy-Tick beim Turn-Ende.
+ *
+ * Income setzt sich zusammen aus Feld-Anteil (1 Gold pro eigenem,
+ * nicht-skeletonisiertem Feld) + Farm-Anteil (3 Gold pro Farm).
+ *
+ * Upkeep ist progressiv: 1. Einheit = 3, 2. = 4, 3. = 5, usw.
+ */
+@Service
+class EconomyService {
+
+    companion object {
+        const val STARTING_GOLD = 6
+        const val FARM_INCOME_PER_ROUND = 3
+        const val FIELD_INCOME_PER_ROUND = 1
+        const val FARM_BASE_COST = 10
+        const val FARM_COST_INCREMENT = 1
+        const val UNIT_PRICE = 5
+    }
+
+    fun computeIncome(player: Player, state: GameState): Int {
+        val ownedFields = state.fields.count { it.owner == player.name && !it.isSkeleton }
+        return ownedFields * FIELD_INCOME_PER_ROUND + player.farms * FARM_INCOME_PER_ROUND
+    }
+
+    fun computeUpkeep(player: Player, state: GameState): Int {
+        val unitCount = state.units.count {
+            it.player == player.name && it.type != UnitType.SKELETON && it.type != UnitType.BASE
+        }
+        return (0 until unitCount).sumOf { 3 + it }
+    }
+
+    fun recomputePlayerStats(state: GameState) {
+        state.players.forEach { player ->
+            player.income = computeIncome(player, state)
+            player.upkeep = computeUpkeep(player, state)
+        }
+    }
+
+    /**
+     * Wirtschafts-Rundenabschluss fuer einen Spieler.
+     * Farm- + Feld-Einkommen gutschreiben, Unterhalt abziehen.
+     * Bei Insolvenz: Gold auf 0, alle lebenden Truppen → SKELETON.
+     */
+    fun applyEconomy(state: GameState, player: Player) {
+        player.gold += computeIncome(player, state)
+        val upkeep = computeUpkeep(player, state)
+        val playerUnits = state.units.filter {
+            it.player == player.name && it.type != UnitType.SKELETON && it.type != UnitType.BASE
+        }
+
+        if (player.gold >= upkeep) {
+            player.gold -= upkeep
+        } else {
+            player.gold = 0
+            playerUnits.forEach { it.type = UnitType.SKELETON }
+        }
+    }
+
+    /**
+     * Kauft eine neue Einheit und platziert sie an (x, y).
+     * Setzt voraus, dass der Controller bereits alle Validierungen
+     * durchgefuehrt hat.
+     */
+    fun buyUnit(
+        state: GameState,
+        playerName: String,
+        type: UnitType,
+        x: Int,
+        y: Int
+    ): GameState = synchronized(state.lock) {
+        val player = state.players.find { it.name == playerName } ?: return state
+
+        state.units.removeIf {
+            it.x == x && it.y == y && it.type == UnitType.SKELETON
+        }
+
+        player.gold -= UNIT_PRICE
+
+        state.units.add(
+            GameUnit(
+                player = playerName,
+                x = x,
+                y = y,
+                type = type,
+                hasMovedThisTurn = true
+            )
+        )
+
+        return state
+    }
+}
